fix: final audit-round-2 findings (auth, test hardening, robustness)
Backend:
- P1 auth: Prompt Library GET endpoints now require auth (were the only
  unauthenticated reads); test flipped to assert 401 without token
- P1: mid-stream Ollama-pull error now has a test (final NDJSON line is {error})
- harden hollow tests: seed idempotency recounts rows (==6); non-Ollama pull
  guard asserts the discriminating message
- constant-time login (secrets.compare_digest)
- document the intentional FK-index downgrade asymmetry in migration a8c4e1f2b3d4

Frontend:
- P1: Ollama streaming-pull tests added (progress bar/statusText, mid-stream
  in-band {error} line, no-provider disabled+hint+guard, emptyName guard)
- fix stale 'stub' comments (md/docx/epub/pdf are all real now)
- dedup i18n (reuse hu.settings.modelPull.noProvider), distinguish
  no-provider vs providers-unreachable, wire AbortSignal in usePullModel,
  promptTemplateReadSchema.uses -> int

Final suites green: web 803, api 534, ai 396; tsc + lint clean.

+++ apps/web/lib/api/types.ts
@@ -378,7 +378,7 @@ export const promptTemplateReadSchema = z.object({
   category: z.string(),
   description: z.string(),
   body: z.string(),
-  uses: z.number(),
+  uses: z.number().int(),
   is_builtin: z.boolean(),
   icon_key: z.string().nullable(),
   created_at: z.string(),
